moved routes and changed paths
