+++ server.js
@@ -8,9 +8,9 @@ var passport = require("passport");
 var Auth0Strategy = require("passport-auth0");
 var flash = require("connect-flash");
 var userInViews = require("./lib/middleware/userInViews");
-var authRouter = require("./routes/auth");
-var indexRouter = require("./routes/index");
-var usersRouter = require("./routes/users");
+var authRouter = require("./api/routes/auth");
+var indexRouter = require("./api/routes/index");
+var usersRouter = require("./api/routes/users");
 var invoicesRouter = require("./api/routes/invoices");
 const bodyParser = require("body-parser");
 const cors = require("cors");
